Refatorações e ajustes no interpretador Delégua.
