+++ fontes/delegua.ts
@@ -6,7 +6,7 @@ import chalk from 'chalk';
 import { Lexador } from './lexador/lexador';
 import { AvaliadorSintatico } from './avaliador-sintatico/avaliador-sintatico';
 import { Resolvedor } from './resolvedor';
-import { Interpretador } from './interpretador';
+import { Interpretador } from './interpretador/interpretador';
 import tiposDeSimbolos from './lexador/tipos-de-simbolos';
 
 import { ExcecaoRetornar } from './excecoes';
